+++ config/config.exs
@@ -120,6 +120,7 @@ config :esbuild,
          js/picker/Picker.tsx
          js/picker/PickerButton.tsx
          js/health/WorkflowHealth.tsx
+         js/health/ProjectHealth.tsx
          editor.worker=monaco-editor/esm/vs/editor/editor.worker.js
          json.worker=monaco-editor/esm/vs/language/json/json.worker.js
          css.worker=monaco-editor/esm/vs/language/css/css.worker.js